acive node panel fully genericized
